1280364: Display version specific links in installer.
Previously we displayed a link to the latest documentation regardless of which
version you were installing. With this change each variant version carries a
unique documentation prefix, and we assume that the path to major sections of
docs we'll be linking to is constant across versions.

It was discovered that the variant and version is not actually known at the
point where we display the prerequisites link, so the choice of which
variant/version now is the first question you will be asked, and preceeds the
specification of hosts. This both solves the problem and feels more correct.